requirements.txt
openai>=1.0.0
flask>=2.0.0
requests>=2.0.0
python-dotenv>=0.21.0
gradio>=4.0.0
